moved isGooglePlayServicesAvailable helper method to WPMainActivity
